# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Research Navigator API - Comprehensive scientific intelligence platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /papers:
    get:
      operationId: listPapers
      tags: [papers]
      summary: List collected papers
      parameters:
        - name: topic
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [citationCount, publicationDate, year]
            default: citationCount
      responses:
        "200":
          description: List of papers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperListResponse"
  /papers/{id}:
    get:
      operationId: getPaper
      tags: [papers]
      summary: Get paper by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paper details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperDetail"
        "404":
          description: Paper not found
...
  /graph/seed/{paperId}:
    get:
      operationId: getCitationGraph
      tags: [graph]
      summary: Get citation graph from seed paper
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
        - name: depth
          in: query
          schema:
            type: integer
            default: 2
            minimum: 1
            maximum: 3
        - name: maxNodes
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Citation graph data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitationGraph"
        "404":
          description: Paper not found
  /graph/paper/{paperId}/summary:
    get:
      operationId: getPaperAiSummary
      tags: [graph]
      summary: Get AI-generated summary for a paper
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI paper summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperSummary"
...
  /gaps/{runId}:
    get:
      operationId: getResearchGaps
      tags: [gaps]
      summary: Get identified research gaps
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Research gaps
          content:
            application/json:
              schema:
                type: object
                properties:
                  gaps:
                    type: array
                    items:
                      $ref: "#/components/schemas/ResearchGap"
                required:
                  - gaps
  /gaps/{runId}/analyze:
    post:
      operationId: analyzeResearchGaps
      tags: [gaps]
      summary: Analyze research gaps for a collection run
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis started
          content:
            application/json:
              schema:
                type: object
                properties:
                  gaps:
                    type: array
                    items:
                      $ref: "#/components/schemas/ResearchGap"
                  message:
                    type: string
                required:
                  - gaps
                  - message
  /proposals/{runId}:
    get:
      operationId: getProposals
      tags: [proposals]
      summary: Get research proposals for a run
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Research proposals
          content:
            application/json:
              schema:
                type: object
                properties:
                  proposals:
                    type: array
                    items:
                      $ref: "#/components/schemas/ResearchProposal"
                required:
                  - proposals
...
components:
  schemas:
...
    Paper:
      type: object
      properties:
        id:
          type: string
        semanticScholarId:
          type: string
          nullable: true
        openAlexId:
          type: string
          nullable: true
        doi:
          type: string
          nullable: true
        title:
          type: string
        abstract:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        citationCount:
          type: integer
        influentialCitationCount:
          type: integer
        venue:
          type: string
          nullable: true
        keywords:
          type: array
          items:
            type: string
        tldr:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        source:
          type: string
      required:
        - id
        - title
        - citationCount
        - influentialCitationCount
        - keywords
        - source
    PaperDetail:
      allOf:
        - $ref: "#/components/schemas/Paper"
        - type: object
          properties:
            authors:
              type: array
              items:
                $ref: "#/components/schemas/Author"
            references:
              type: array
              items:
                $ref: "#/components/schemas/Paper"
            citations:
              type: array
              items:
                $ref: "#/components/schemas/Paper"
          required:
            - authors
            - references
            - citations
    PaperListResponse:
      type: object
      properties:
        papers:
          type: array
          items:
            $ref: "#/components/schemas/Paper"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - papers
        - total
        - limit
        - offset
    Author:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        affiliations:
          type: array
          items:
            type: string
        citationCount:
          type: integer
        paperCount:
          type: integer
        hIndex:
          type: integer
      required:
        - id
        - name
        - affiliations
        - citationCount
        - paperCount
        - hIndex
...
    CitationGraphNode:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        year:
          type: integer
          nullable: true
        citationCount:
          type: integer
        influentialCitationCount:
          type: integer
        depth:
          type: integer
        isHub:
          type: boolean
        isBridge:
          type: boolean
        cluster:
          type: string
          nullable: true
        authors:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - citationCount
        - influentialCitationCount
        - depth
        - isHub
        - isBridge
        - authors
    CitationGraphEdge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        isInfluential:
          type: boolean
      required:
        - source
        - target
        - isInfluential
    ResearchLineage:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        paperIds:
          type: array
          items:
            type: string
        description:
          type: string
      required:
        - id
        - name
        - paperIds
        - description
    CitationGraph:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/CitationGraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/CitationGraphEdge"
        seedPaperId:
          type: string
        totalNodes:
          type: integer
        totalEdges:
          type: integer
        lineages:
          type: array
          items:
            $ref: "#/components/schemas/ResearchLineage"
      required:
        - nodes
        - edges
        - seedPaperId
        - totalNodes
        - totalEdges
        - lineages
    PaperSummary:
      type: object
      properties:
        paperId:
          type: string
        title:
          type: string
        summary:
          type: string
        keyContributions:
          type: array
          items:
            type: string
        methodology:
          type: string
        impact:
          type: string
      required:
        - paperId
        - title
        - summary
        - keyContributions
        - methodology
        - impact
...
    ResearchGap:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        evidenceType:
          type: string
          enum: [topic_modeling, citation_network, method_problem_matrix]
        supportingPaperIds:
          type: array
          items:
            type: string
        noveltyScore:
          type: number
        impactScore:
          type: number
        feasibilityScore:
          type: number
      required:
        - id
        - title
        - description
        - evidenceType
        - supportingPaperIds
        - noveltyScore
        - impactScore
        - feasibilityScore
    ResearchProposal:
      type: object
      properties:
        id:
          type: string
        gapId:
          type: string
        title:
          type: string
        motivation:
          type: string
        researchQuestions:
          type: array
          items:
            type: string
        methodology:
          type: string
        expectedContributions:
          type: array
          items:
            type: string
        challenges:
          type: array
          items:
            type: string
        noveltyScore:
          type: number
        noveltyExplanation:
          type: string
          nullable: true
        supportingPaperIds:
          type: array
          items:
            type: string
        rawText:
          type: string
          nullable: true
      required:
        - id
        - gapId
        - title
        - motivation
        - researchQuestions
        - methodology
        - expectedContributions
        - challenges
        - noveltyScore
        - supportingPaperIds
...